Add a file watcher and config reference system
This allows creating configuration references that automatically reload,
and deriving value references based on those configurations to keep
values up to date when the configuration changes.

We also extend the reactive system to allow for transactional
subscriptions, so that if any single value fails to be deserialized the
entire configuration will not reload.